Move functor-based matrix code to a separate file
